requirements.txt
fastapi==0.66.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.0
h11==0.12.0
Mako==1.1.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
promise==2.3
pydantic==1.8.2
python-dateutil==2.8.1
python-dotenv==0.18.0
regex==2021.7.1
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
pymongo==4.1.1
dnspython==2.2.1